fix(ui): normalize neural orb math for Kotlin

+++ app/src/main/java/com/niko/assistant/ui/NikoHero.kt
@@ -134,19 +134,18 @@ internal fun NikoHero(
                 val steps = 112
                 for (index in 0..steps) {
                     val angle = (index.toFloat() / steps) * (PI * 2).toFloat()
-                    val waveA = sin(angle * lobesA + phase * PI.toFloat() * 2f + phaseOffset)
-                    val waveB = cos(angle * lobesB - phase * PI.toFloat() * 1.65f - phaseOffset)
-                    val waveC = sin(angle * 2f + phase * PI.toFloat() * 0.8f)
+                    val waveA = sin((angle * lobesA + phase * PI.toFloat() * 2f + phaseOffset).toDouble()).toFloat()
+                    val waveB = cos((angle * lobesB - phase * PI.toFloat() * 1.65f - phaseOffset).toDouble()).toFloat()
+                    val waveC = sin((angle * 2f + phase * PI.toFloat() * 0.8f).toDouble()).toFloat()
                     val r = radius + distortion * (waveA * 0.55f + waveB * 0.30f + waveC * 0.15f)
-                    val x = center.x + cos(angle) * r
-                    val y = center.y + sin(angle) * r
+                    val x = center.x + cos(angle.toDouble()).toFloat() * r
+                    val y = center.y + sin(angle.toDouble()).toFloat() * r
                     if (index == 0) path.moveTo(x, y) else path.lineTo(x, y)
                 }
                 path.close()
                 return path
             }
 
-            // Atmospheric halo: the visual presence is larger than the actual core.
             drawCircle(
                 brush = Brush.radialGradient(
                     colors = listOf(
@@ -161,7 +160,6 @@ internal fun NikoHero(
                 center = center,
             )
 
-            // Listening expands in clean acoustic ripples instead of showing a generic spinner.
             if (state == NikoVisualState.LISTENING) {
                 repeat(4) { index ->
                     val progress = (phase * 1.45f + index * 0.25f) % 1f
@@ -174,7 +172,6 @@ internal fun NikoHero(
                 }
             }
 
-            // Thin orbital arcs add depth without turning the UI into a sci-fi HUD overload.
             repeat(3) { index ->
                 val radius = unit * (0.335f + index * 0.035f)
                 val spin = phase * 360f * (if (index % 2 == 0) 1f else -0.72f) + index * 71f
@@ -191,17 +188,16 @@ internal fun NikoHero(
                 }
             }
 
-            // A small particle field makes thinking/speaking feel alive but remains deterministic.
             val particleCount = if (state == NikoVisualState.IDLE) 8 else 16
             repeat(particleCount) { index ->
                 val baseAngle = (index.toFloat() / particleCount) * (PI * 2).toFloat()
                 val direction = if (index % 2 == 0) 1f else -1f
                 val angle = baseAngle + phase * direction * (0.55f + index % 3 * 0.08f) * (PI * 2).toFloat()
                 val orbit = unit * (0.34f + (index % 4) * 0.025f)
-                val wobble = unit * 0.018f * sin(phase * PI.toFloat() * 4f + index)
+                val wobble = unit * 0.018f * sin((phase * PI.toFloat() * 4f + index).toDouble()).toFloat()
                 val point = Offset(
-                    center.x + cos(angle) * (orbit + wobble),
-                    center.y + sin(angle) * (orbit * 0.72f + wobble),
+                    center.x + cos(angle.toDouble()).toFloat() * (orbit + wobble),
+                    center.y + sin(angle.toDouble()).toFloat() * (orbit * 0.72f + wobble),
                 )
                 drawCircle(
                     color = (if (index % 3 == 0) accentA else accentB).copy(alpha = 0.20f + energy * 0.38f),
@@ -222,7 +218,6 @@ internal fun NikoHero(
             val body = organicPath(baseRadius, distortion, 1.05f, 4, 7)
             val core = organicPath(baseRadius * 0.67f, distortion * 0.42f, 2.1f, 3, 6)
 
-            // Outer chromatic membrane.
             drawPath(
                 path = outer,
                 brush = Brush.sweepGradient(
@@ -238,7 +233,6 @@ internal fun NikoHero(
                 alpha = 0.58f + energy * 0.28f,
             )
 
-            // Dark liquid-glass body with a colored internal light source.
             drawPath(
                 path = body,
                 brush = Brush.radialGradient(
@@ -266,7 +260,6 @@ internal fun NikoHero(
                 alpha = 0.88f,
             )
 
-            // Luminous neural core.
             drawPath(
                 path = core,
                 brush = Brush.radialGradient(
@@ -282,7 +275,6 @@ internal fun NikoHero(
                 alpha = 0.62f + energy * 0.30f,
             )
 
-            // Specular highlight sells the glass material on OLED and LCD screens alike.
             drawCircle(
                 brush = Brush.radialGradient(
                     colors = listOf(Color.White.copy(alpha = 0.28f), Color.Transparent),
@@ -293,7 +285,6 @@ internal fun NikoHero(
                 center = Offset(center.x - unit * 0.09f, center.y - unit * 0.11f),
             )
 
-            // Minimal NIKO monogram: recognizable without putting text inside Canvas.
             val markColor = Color.White.copy(alpha = 0.88f)
             val markWidth = unit * 0.010f
             val markLeft = center.x - unit * 0.045f
@@ -304,15 +295,20 @@ internal fun NikoHero(
             drawLine(markColor, Offset(markLeft, markTop), Offset(markRight, markBottom), markWidth, StrokeCap.Round)
             drawLine(markColor, Offset(markRight, markBottom), Offset(markRight, markTop), markWidth, StrokeCap.Round)
 
-            // Speaking energy bars sit on the orbit, not on a cartoon mouth.
             if (state == NikoVisualState.SPEAKING) {
                 repeat(20) { index ->
                     val angle = (index / 20f) * (PI * 2).toFloat()
-                    val wave = abs(sin(phase * PI.toFloat() * 7f + index * 0.83f))
+                    val wave = abs(sin((phase * PI.toFloat() * 7f + index * 0.83f).toDouble()).toFloat())
                     val inner = unit * 0.305f
                     val outerRadius = inner + unit * (0.018f + wave * 0.050f)
-                    val start = Offset(center.x + cos(angle) * inner, center.y + sin(angle) * inner)
-                    val end = Offset(center.x + cos(angle) * outerRadius, center.y + sin(angle) * outerRadius)
+                    val start = Offset(
+                        center.x + cos(angle.toDouble()).toFloat() * inner,
+                        center.y + sin(angle.toDouble()).toFloat() * inner,
+                    )
+                    val end = Offset(
+                        center.x + cos(angle.toDouble()).toFloat() * outerRadius,
+                        center.y + sin(angle.toDouble()).toFloat() * outerRadius,
+                    )
                     drawLine(
                         color = (if (index % 2 == 0) magenta else cyan).copy(alpha = 0.46f + wave * 0.40f),
                         start = start,
